Clean up strategic programme logic
