+++ controllers/strategic-investments/index.js
@@ -1,12 +1,12 @@
 'use strict';
 const path = require('path');
 const express = require('express');
-const { get, set, startsWith } = require('lodash');
+const get = require('lodash/get');
+const set = require('lodash/set');
 
 const {
     injectBreadcrumbs,
     injectListingContent,
-    injectStrategicProgramme,
     setCommonLocals
 } = require('../../common/inject-content');
 const contentApi = require('../../common/content-api');
@@ -40,51 +40,67 @@ router.get('/', injectListingContent, async function(req, res, next) {
     }
 });
 
+async function injectStrategicProgramme(req, res, next) {
+    try {
+        // Assumes a parameter of :slug and :childPageSlug? in the request
+        const { slug, childPageSlug } = req.params;
+        if (slug) {
+            const querySlug = childPageSlug ? `${slug}/${childPageSlug}` : slug;
+
+            const entry = await contentApi.getStrategicProgrammes({
+                slug: querySlug,
+                locale: req.i18n.getLocale(),
+                requestParams: req.query
+            });
+
+            res.locals.strategicProgramme = entry;
+            setCommonLocals({ res, entry });
+        }
+        next();
+    } catch (error) {
+        if (error.statusCode >= 500) {
+            next(error);
+        } else {
+            next();
+        }
+    }
+}
+
 router.get('/:slug/:childPageSlug?', injectStrategicProgramme, function(
     req,
     res,
     next
 ) {
-    const { currentPath, strategicProgramme } = res.locals;
-    /* Only render if not using an external path */
-    if (
-        strategicProgramme &&
-        startsWith(currentPath, strategicProgramme.linkUrl)
-    ) {
-        // Render a plain content page if specified
-        if (get(strategicProgramme, 'entryType') === 'contentPage') {
-            setCommonLocals({ res, entry: strategicProgramme });
-            set(
-                res.locals,
-                'content.flexibleContent',
-                strategicProgramme.content
-            );
+    const { strategicProgramme } = res.locals;
+    // Render a plain content page if specified
+    if (get(strategicProgramme, 'entryType') === 'contentPage') {
+        setCommonLocals({ res, entry: strategicProgramme });
+        set(res.locals, 'content.flexibleContent', strategicProgramme.content);
 
-            const breadcrumbs = strategicProgramme.parent
-                ? res.locals.breadcrumbs.concat([
-                      {
-                          label: strategicProgramme.parent.title,
-                          url: strategicProgramme.parent.linkUrl
-                      },
-                      { label: res.locals.title }
-                  ])
-                : res.locals.breadcrumbs.concat([{ label: res.locals.title }]);
+        const breadcrumbs = strategicProgramme.parent
+            ? res.locals.breadcrumbs.concat([
+                  {
+                      label: strategicProgramme.parent.title,
+                      url: strategicProgramme.parent.linkUrl
+                  },
+                  { label: res.locals.title }
+              ])
+            : res.locals.breadcrumbs.concat([{ label: res.locals.title }]);
 
-            res.render(
-                path.resolve(__dirname, '../common/views/flexible-content'),
-                {
-                    breadcrumbs: breadcrumbs,
-                    flexibleContent: strategicProgramme.content
-                }
-            );
-        } else {
-            // Render a standard Strategic page
-            res.render(path.resolve(__dirname, './views/strategic-programme'), {
-                breadcrumbs: res.locals.breadcrumbs.concat({
-                    label: res.locals.title
-                })
-            });
-        }
+        res.render(
+            path.resolve(__dirname, '../common/views/flexible-content'),
+            {
+                breadcrumbs: breadcrumbs,
+                flexibleContent: strategicProgramme.content
+            }
+        );
+    } else if (strategicProgramme) {
+        // Render a standard Strategic page
+        res.render(path.resolve(__dirname, './views/strategic-programme'), {
+            breadcrumbs: res.locals.breadcrumbs.concat({
+                label: res.locals.title
+            })
+        });
     } else {
         next();
     }
